Updated the coins page + the not found page + organized folders and files
