+++ src/App.css
@@ -5,6 +5,17 @@ body {
   color: antiquewhite;
 }
 
+a:link {
+  color: white;
+  text-decoration: none;
+}
+
+a:visited {
+  color: white;
+  text-decoration: none;
+}
+
+
 /* NAVBAR CSS */
 .nav-btn {
   margin-left: 20px;
@@ -132,7 +143,7 @@ body {
   height: 50px;
   border-radius: 4px;
   border: none;
-  background-image: linear-gradient(-225deg,#ac32e4 0%, #7918f2 48%, #4801ff 100% );
+  background-image: linear-gradient(-225deg,#8426af 0%, #7918f2 48%, #3a0968 100% );
   color: #e2e2e2;
 }
 
@@ -140,6 +151,48 @@ body {
   color: #e2e2e2;
 }
 
+/* COIN INFO PAGE CSS */
+.coinInfo-main {
+  margin: 100px 50% 0 21%;
+  color: white;
+}
+
+.coinInfo-h1 {
+  text-align: center;
+}
+
+.coinInfo-card {
+  padding: 10%;
+}
+
+.coinInfo-h3 {
+  margin: 3% 3% 3% 0%;
+}
+
+.coinInfo-hr {
+  width: 100%;
+  background-color: rgb(250, 29, 239);
+  border-width: 4px;
+  margin-top: 40px;
+}
+
+.coinInfo-img {
+  width: 10rem;
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 20px;
+}
+
+table{
+  border: 1px solid;
+}
+th {
+  border: 1px solid;
+}
+td {
+  border: 1px solid;
+}
 
 /* ABOUT PAGE CSS */
 .about-card {
@@ -165,4 +218,13 @@ body {
   background-color: rgb(226, 43, 217);
   border-width: 2px;
   margin-top: 40px;
+}
+
+/* PAGE NOT FOUND CSS */
+.notfound-main {
+  margin: auto;
+  margin-top: 20%;
+  width: 50%;
+  text-align: center;
+  color: red;
 }
