+++ src/NotFound.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function NotFound() {
+  return (
+    <div className='notfound-main'>
+        <h1 className='notfound-h1'>The page you are looking for was not found :(</h1>
+    </div>
+  )
+}
+
+export default NotFound
